Workaround yt-dlp not downloading uploads, ignore georestrict errors

+++ sns-archive/src/youtube/mod.rs
@@ -1,8 +1,12 @@
 use std::ffi::OsString;
 use std::fmt::Display;
-use std::fs;
+use std::io::{Read, Write};
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
+use std::{fs, mem};
+
+use once_cell::sync::Lazy;
+use regex::Regex;
 
 use crate::config::youtube::{YTChannel, YoutubeConfig};
 
@@ -52,19 +56,27 @@ pub fn download(config: YoutubeConfig) -> Result<(), YTError> {
             new_channel,
             &config.archive_path,
         );
-        let output = Command::new("yt-dlp")
+        let mut child = Command::new("yt-dlp")
             .stdout(Stdio::inherit())
-            .stderr(Stdio::inherit())
+            .stderr(Stdio::piped())
             .args(&args)
-            .output()
+            .spawn()
             .expect("Failed to execute command");
-        if !output.status.success() {
-            eprintln!("{}", output.status);
+
+        // Tee stderr to in memory buffer to inspect any errors
+        let child_err = mem::take(&mut child.stderr).unwrap();
+        let stderr = std::thread::spawn(move || tee(child_err, std::io::stderr()).unwrap())
+            .join()
+            .unwrap();
+
+        let status = child.wait().unwrap();
+        if !status.success() && !error_is_geo_restrict(&String::from_utf8(stderr).unwrap()) {
+            eprintln!("{}", status);
             errored_channels.push(channel);
             continue;
         }
 
-        if new_channel {
+        if new_channel && directory_exists(&tmp_dir) {
             fs::rename(&tmp_dir, &dir).unwrap_or_else(|_| {
                 panic!("Could not rename directory {:?} to {:?}", &tmp_dir, &dir)
             });
@@ -132,6 +144,23 @@ fn generate_cmd_args(
     args
 }
 
+fn tee(mut stream: impl Read, mut output_2: impl Write) -> std::io::Result<Vec<u8>> {
+    let mut buf = [0u8; 1024];
+    let mut output = Vec::new();
+    loop {
+        let num_read = stream.read(&mut buf)?;
+        if num_read == 0 {
+            break;
+        }
+
+        let buf = &buf[..num_read];
+        output.write_all(buf)?;
+        output_2.write_all(buf)?;
+    }
+
+    Ok(output)
+}
+
 fn directory_exists(path: &impl AsRef<Path>) -> bool {
     match fs::metadata(path) {
         Err(_) => false,
@@ -140,5 +169,20 @@ fn directory_exists(path: &impl AsRef<Path>) -> bool {
 }
 
 fn channel_id_to_url(channel_id: &str) -> String {
-    format!("https://www.youtube.com/channel/{}", channel_id)
+    format!(
+        "https://www.youtube.com/playlist?list=UU{}",
+        &channel_id[2..]
+    )
+}
+
+fn error_is_geo_restrict(log: &str) -> bool {
+    static ERROR_RE: Lazy<Regex> = Lazy::new(|| Regex::new(r"^ERROR:").unwrap());
+    static ERROR_MSG_RE: Lazy<Regex> = Lazy::new(|| {
+        Regex::new(r"The uploader has not made this video available in your country").unwrap()
+    });
+    0 == log
+        .lines()
+        .filter(|line| ERROR_RE.find(line.trim()).is_some())
+        .filter(|line| ERROR_MSG_RE.find(line).is_none())
+        .count()
 }
